Adds a comma to the 'goals' example in README.rst
The current example for adding a goal in the README does not give a
tuple. Python will evaluate the example as a single string and will iterate over
the string, creating a new goal for each letter. Adding a comma after
the string turns it into a tuple which means that only one goal will be
created.